+++ packages/asc-ui/src/components/GlobalStyle/GlobalStyle.ts
@@ -10,11 +10,17 @@ const GlobalStyle = createGlobalStyle`
     font-family: ${getValueFromTheme('typography.fontFamily')};
   }
 
+  /* Display reset for older browsers that don't support certain HTML5 elements */
+  article, aside, details, figcaption, figure, footer, header, hgroup, menu, nav, section {
+      display: block;
+  }
+
   [aria-hidden="true"] {
     display: none !important;
   }
   
-  * {
+  /* Use border-box sizing by default for all elements */
+  *, *::before, *::after {
     box-sizing: border-box;
   }